перевёл хар-ки на float

+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Common/IHeatResistable.cs
@@ -4,7 +4,7 @@ namespace Game.Gameplay.SurvivalMechanics
 {
     public interface IHeatResistable
     {
-        int HeatResistance { get; set; }
-        event Action<int> HeatResistanceChanged;
+        float HeatResistance { get; set; }
+        event Action<float> HeatResistanceChanged;
     }
 }
