+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Exhaustion/ExhaustionConfig.cs
@@ -5,8 +5,8 @@ namespace Game.Gameplay.SurvivalMeсhanics.Exhaustion
     [CreateAssetMenu(fileName = "Exhaustion Config", menuName = "Configs/Exhaustion")]
     public class ExhaustionConfig : ScriptableObject
     {
-        [SerializeField, Tooltip("Per Second")] private int _damage;
+        [SerializeField, Tooltip("Per Second")] private float _damage;
         
-        public int Damage => _damage;
+        public float Damage => _damage;
     }
 }
