+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Exhaustion/ExhaustionController.cs
@@ -12,12 +12,12 @@ namespace Game.Gameplay.SurvivalMeсhanics.Exhaustion
             _metricsModel = metricsModel;
         }
 
-        protected override void SubscribeOnMetric(System.Action<int> handler)
+        protected override void SubscribeOnMetric(System.Action<float> handler)
         {
             _metricsModel.HungerChanged += handler;
         }
 
-        protected override void UnsubscribeFromMetric(System.Action<int> handler)
+        protected override void UnsubscribeFromMetric(System.Action<float> handler)
         {
             _metricsModel.HungerChanged -= handler;
         }
