+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Frostbite/FrostbiteConfig.cs
@@ -5,8 +5,8 @@ namespace Game.Gameplay.SurvivalMeсhanics.Frostbite
     [CreateAssetMenu(fileName = "Frostbite Config", menuName = "Configs/Frostbite")]
     public class FrostbiteConfig : ScriptableObject
     {
-        [SerializeField, Tooltip("Per Second")] private int _damage;
+        [SerializeField, Tooltip("Per Second")] private float _damage;
         
-        public int Damage => _damage;
+        public float Damage => _damage;
     }
 }
