+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Hunger/HungerConfig.cs
@@ -5,8 +5,8 @@ namespace Game.Gameplay.SurvivalMeсhanics.Hunger
     [CreateAssetMenu(fileName = "Hunger Config", menuName = "Configs/Hunger")]
     public class HungerConfig : ScriptableObject
     {
-        [SerializeField] private int[] _consumptions;
+        [SerializeField] private float[] _consumptions;
 
-        public int GetConsumption(int level) => _consumptions[level - 1];
+        public float GetConsumption(int level) => _consumptions[level - 1];
     }
 }
