+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/PlayerMetrics/PlayerMetricsConfig.cs
@@ -5,16 +5,16 @@ namespace Game.Gameplay.Survival_Metrics.Configs
     [CreateAssetMenu(fileName = "Player Metrics Config", menuName = "Configs/PlayerMetrics")]
     public class PlayerMetricsConfig : ScriptableObject
     {
-        [SerializeField] private int _health;
-        [SerializeField] private int _heatResistance;
-        [SerializeField] private int _hunger;
-        [SerializeField] private int _thirst;
-        [SerializeField] private int _endurance;
+        [SerializeField] private float _health;
+        [SerializeField] private float _heatResistance;
+        [SerializeField] private float _hunger;
+        [SerializeField] private float _thirst;
+        [SerializeField] private float _endurance;
         
-        public int Health => _health;
-        public int HeatResistance => _heatResistance;
-        public int Hunger => _hunger;
-        public int Thirst => _thirst;
-        public int Endurance => _endurance;
+        public float Health => _health;
+        public float HeatResistance => _heatResistance;
+        public float Hunger => _hunger;
+        public float Thirst => _thirst;
+        public float Endurance => _endurance;
     }
 }
