+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Thirst/ThirstConfig.cs
@@ -5,8 +5,8 @@ namespace Game.Gameplay.SurvivalMeсhanics.Thirst
     [CreateAssetMenu(fileName = "Thirst Config", menuName = "Configs/Thirst")]
     public class ThirstConfig : ScriptableObject
     {
-        [SerializeField] private int[] _consumptions;
+        [SerializeField] private float[] _consumptions;
 
-        public int GetConsumption(int level) => _consumptions[level - 1];
+        public float GetConsumption(int level) => _consumptions[level - 1];
     }
 }
